Start Policy

+++ app/Http/Controllers/GroupController.php
@@ -79,14 +79,15 @@ class GroupController extends Controller
     {
         $group = Group::find($id);
 
-        if (Gate::denies('can-access-group', $group)){
-            return redirect()->route('home')->with('danger','Vous n\'appartenez pas à ce groupe. Par conséquent vous ne pouvez pas y acceder');
+        if (Auth::user()->can('view', $group)) {
+            $members = UserGroup::where('group_id', $group->id)->get();
+            return view('group.show', [
+                'group' => $group,
+                'members' => $members
+            ]);
         }
-        $members = UserGroup::where('group_id', $group->id)->get();
-        return view('group.show',[
-            'group'=>$group,
-            'members'=>$members
-        ]);
+
+        return redirect()->route('home')->with('danger', 'Vous n\'appartenez pas à ce groupe. Par conséquent vous ne pouvez pas y acceder');
     }
 
     /**
@@ -99,18 +100,18 @@ class GroupController extends Controller
     {
         $group = Group::find($id);
 
-        if (Gate::denies('can-access-group', $group)){
-            return redirect()->route('home')->with('danger','Vous n\'êtes pas propriétaire du groupe, vous n\'avez pas accès à cette page');
+        if (Auth::user()->can('update', $group)) {
+            $members = UserGroup::where('group_id', $group->id)->where('user_id', '!=', $group->user_id)->get();
+            //members dans le 2nd where on exclut du groupe la personne propriétaire et on récupère les autres
+
+
+            return view('group.edit', [
+                'group' => $group,
+                'members' => $members
+            ]);
         }
 
-        $members = UserGroup::where('group_id', $group->id)->where('user_id', '!=', $group->user_id)->get();
-        //members dans le 2nd where on exclut du groupe la personne propriétaire et on récupère les autres
-
-
-        return view('group.edit', [
-            'group' => $group,
-            'members' => $members
-        ]);
+        return redirect()->route('home')->with('danger', 'Vous n\'êtes pas propriétaire du groupe, vous n\'avez pas accès à cette page');
 
     }
 
@@ -125,17 +126,17 @@ class GroupController extends Controller
     {
         $group = Group::find($id);
 
-        if (Gate::denies('can-edit-group', $group)){
-            return redirect()->route('home')->with('danger','Vous n\'êtes pas propriétaire du groupe, vous ne pouvez pas modifier ses informations');
+        if (Auth::user()->can('update', $group)) {
+            if ($request->input('name') != null) {
+                $group->name = $request->input('name');
+            }
+
+            $group->save();
+
+            return redirect()->route('group.show', $group->id)->with('success', 'Les informations du groupe ont été modifiées avec succès');
         }
 
-        if ($request->input('name') != null) {
-            $group->name = $request->input('name');
-        }
-
-        $group->save();
-
-        return redirect()->route('group.show', $group->id)->with('success','Les informations du groupe ont été modifiées avec succès');
+        return redirect()->route('home')->with('danger', 'Vous n\'êtes pas propriétaire du groupe, vous ne pouvez pas modifier ses informations');
     }
 
     /**
@@ -148,43 +149,43 @@ class GroupController extends Controller
     {
         $group = Group::find($id);
 
-        if (Gate::denies('can-edit-group', $group)){
-            return redirect()->route('home')->with('danger','Vous n\'êtes pas propriétaire du groupe, vous ne pouvez pas supprimer le groupe');
-        }
-
-        $members = UserGroup::where('group_id', $group->id)->get();
-        //Try if userGroups array is empty
-        if (count($members) > 0) {
-            // if isn't empty, delete item
-            foreach ($members as $member)
-                $member->delete();
-        }
-
-        //delete the share directories from the group
-        $directories = ShareDirectory::where('group_id', $group->id)->get();
-
-        if (count($directories) > 0){
-            foreach ($directories as $directory) {
-                $directory->delete();
+        if (Auth::user()->can('delete', $group)) {
+            $members = UserGroup::where('group_id', $group->id)->get();
+            //Try if userGroups array is empty
+            if (count($members) > 0) {
+                // if isn't empty, delete item
+                foreach ($members as $member)
+                    $member->delete();
             }
-        }
 
-        // delete the shares and the policies of the shares
-        if (count($group->shares) >0){
-            foreach ($group->shares as $share){
-                $policies = ShareGroupPolicies::where('shareGroup_id', $share->id)->get();
-                if (count($policies) > 0){
-                    foreach ($policies as $policy){
-                        $policy->delete();
-                    }
+            //delete the share directories from the group
+            $directories = ShareDirectory::where('group_id', $group->id)->get();
+
+            if (count($directories) > 0) {
+                foreach ($directories as $directory) {
+                    $directory->delete();
                 }
-                $share->delete();
             }
-        }
 
-        // and delete the group
-        $group->delete();
-        return redirect()->route('group.index')->with('success','Groupe supprimé avec succès');
+            // delete the shares and the policies of the shares
+            if (count($group->shares) > 0) {
+                foreach ($group->shares as $share) {
+                    $policies = ShareGroupPolicies::where('shareGroup_id', $share->id)->get();
+                    if (count($policies) > 0) {
+                        foreach ($policies as $policy) {
+                            $policy->delete();
+                        }
+                    }
+                    $share->delete();
+                }
+            }
+
+            // and delete the group
+            $group->delete();
+            return redirect()->route('group.index')->with('success', 'Groupe supprimé avec succès');
+        }
+        return redirect()->route('home')->with('danger', 'Vous n\'êtes pas propriétaire du groupe, vous ne pouvez pas supprimer le groupe');
+
     }
 
     /**
@@ -195,19 +196,20 @@ class GroupController extends Controller
      */
     public function exit($id)
     {
-        $userGroups = UserGroup::where('user_id',Auth::user()->id)->where('group_id', $id)->get();
+        $userGroups = UserGroup::where('user_id', Auth::user()->id)->where('group_id', $id)->get();
 
-        if (Gate::denies('is-member-group', $userGroups[0]->group)){
-            return redirect()->route('home')->with('danger','Vous ne pouvez pas quitter ce groupe');
+        //@TODO à voir la policy ici
+        if (Gate::denies('is-member-group', $userGroups[0]->group)) {
+            return redirect()->route('home')->with('danger', 'Vous ne pouvez pas quitter ce groupe');
         }
 
         $shares = $userGroups[0]->user->sharesGroups;
 
-        foreach ($shares as $share){
-            if ($share->group_id == $id){
+        foreach ($shares as $share) {
+            if ($share->group_id == $id) {
                 $policies = ShareGroupPolicies::where('shareGroup_id', $share->id)->get();
 
-                if (count($policies) > 0){
+                if (count($policies) > 0) {
                     foreach ($policies as $policy) {
                         $policy->delete();
                     }
@@ -219,16 +221,16 @@ class GroupController extends Controller
 
         $directory = ShareDirectory::where('user_id', Auth::user()->id)->where('group_id', $id)->get();
 
-        if (count($directory) > 0){
+        if (count($directory) > 0) {
             $directory[0]->delete();
         }
 
 
-        NotificationController::notificationAuto("Vous venez de quitter le groupe ".$userGroups[0]->group->name,"Bonjour, voici un mail vous informant que vous venez de quitter le groupe ".$userGroups[0]->group->name.".");
+        NotificationController::notificationAuto("Vous venez de quitter le groupe " . $userGroups[0]->group->name, "Bonjour, voici un mail vous informant que vous venez de quitter le groupe " . $userGroups[0]->group->name . ".");
 
         $userGroups[0]->delete();
 
-        return redirect()->route('group.index')->with('success','Vous avez bien quitter le groupe.');
+        return redirect()->route('group.index')->with('success', 'Vous avez bien quitter le groupe.');
     }
 
     /**
@@ -241,19 +243,19 @@ class GroupController extends Controller
     {
         $group = Group::find($id);
 
-        if (Gate::denies('can-edit-group', $group)){
-            return redirect()->route('home')->with('danger','Vous ne pouvez pas exclure cette personne');
+        if (Gate::denies('can-edit-group', $group)) {
+            return redirect()->route('home')->with('danger', 'Vous ne pouvez pas exclure cette personne');
         }
 
         $member = UserGroup::where('group_id', $id)->where('user_id', $user_id)->get();
 
         $shares = $member[0]->user->sharesGroups;
 
-        foreach ($shares as $share){
-            if ($share->group_id == $group->id){
+        foreach ($shares as $share) {
+            if ($share->group_id == $group->id) {
                 $policies = ShareGroupPolicies::where('shareGroup_id', $share->id)->get();
 
-                if (count($policies) > 0){
+                if (count($policies) > 0) {
                     foreach ($policies as $policy) {
                         $policy->delete();
                     }
@@ -263,19 +265,19 @@ class GroupController extends Controller
             }
         }
 
-        $directory = ShareDirectory::where('user_id', $user_id )->where('group_id', $group->id)->get();
+        $directory = ShareDirectory::where('user_id', $user_id)->where('group_id', $group->id)->get();
 
-        if (count($directory) > 0){
+        if (count($directory) > 0) {
             $directory[0]->delete();
         }
 
 
-        NotificationController::notificationAutoKick("Vous venez d'être exclue du groupe ".$group->name,"Bonjour, voici un mail vous informant que vous venez d'être exclue du groupe ".$group->name.".",$member[0]->user->id);
+        NotificationController::notificationAutoKick("Vous venez d'être exclue du groupe " . $group->name, "Bonjour, voici un mail vous informant que vous venez d'être exclue du groupe " . $group->name . ".", $member[0]->user->id);
 
         $member[0]->delete();
 
 
-        return redirect()->route('group.show', $group->id)->with('success','La personne a bien été exclue !');
+        return redirect()->route('group.show', $group->id)->with('success', 'La personne a bien été exclue !');
     }
 
     /**
@@ -284,18 +286,18 @@ class GroupController extends Controller
      *
      * display the view with all shares of the group
      */
-    public function share($id){
+    public function share($id)
+    {
 
         $group = Group::find($id);
 
 
-
-        if (Gate::denies('can-access-group', $group)){
-            return redirect()->route('home')->with('danger','Vous n\'appartenez pas à ce groupe. Par conséquent vous ne pouvez pas y acceder');
+        if (Gate::denies('can-access-group', $group)) {
+            return redirect()->route('home')->with('danger', 'Vous n\'appartenez pas à ce groupe. Par conséquent vous ne pouvez pas y acceder');
         }
 
-        return view('group.share.share',[
-            'group'=>$group
+        return view('group.share.share', [
+            'group' => $group
         ]);
     }
